functions/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.12.0
cryptography==45.0.5
deprecation==2.1.0
flask==3.1.1
functions-framework==3.8.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-functions==1.20.4
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
watchdog==6.0.0
werkzeug==3.1.3
